requirements.txt
pytesseract~=0.3.8
Pillow~=8.4.0
Flask~=2.0.2
beautifulsoup4~=4.10.0
requests~=2.26.0